Uplift fast rejection to new solver
